// frontend/src/types/api.ts
export interface ProductBadge {
  slug: string;
  display_text: string;
  display_color: string;
}

export interface ProductSummary {
  id: number;
  name: string;
  slug: string;
  basePrice: string | number;
  brandName?: string;
  mainImageUrl: string | null;
  badgeId?: number | null;
  colors?: string[];
  badge?: ProductBadge; // Giữ lại nếu cần map thêm ở frontend hoặc backend trả về object badge
}